+++ src/dataTypes/hsNumber.js
@@ -1,3 +1,10 @@
+/*
+ * vim: set tw=100 et ts=4 sw=4 si fileencoding=utf-8:
+ * © 2022 WideSky.Cloud Pty Ltd
+ * SPDX-License-Identifier: MIT
+ */
+"use strict";
+
 /**
  * Project Haystack ZINC number regular expression.  Not perfect, but it'll do
  * for now.